pyproject.toml
[project]
name = "celery-example"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4.0",
    "django>=5.1.7",
    "django-celery-beat>=2.7.0",
    "django-celery-results>=2.5.1",
    "django-debug-toolbar>=6.0.0",
    "django-filter>=25.1",
    "django-stubs[compatible-mypy]>=5.2.2",
    "djangorestframework>=3.15.2",
    "drf-nested-routers>=0.94.2",
    "drf-spectacular>=0.28.0",
    "flower>=2.0.1",
    "psycopg[binary]>=3.2.5",
    "redis>=5.2.1",
]
